requirements.txt
appdirs==1.4.4
attrs==20.3.0
beautifulsoup4==4.9.3
black==20.8b1
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
fake-useragent==0.1.11
google-auth==1.23.0
google-auth-oauthlib==0.4.2
gspread==3.6.0
idna==2.10
iniconfig==1.1.1
loguru==0.5.3
lxml==4.6.1
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.4
parse==1.18.0
pathspec==0.8.1
pluggy==0.13.1
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==7.0.4
pyparsing==2.4.7
pyppeteer==0.2.2
pyquery==1.4.3
pytest==6.1.2
python-dotenv==0.15.0
regex==2020.11.13
requests==2.25.0
requests-html==0.10.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
toml==0.10.2
tqdm==4.53.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
w3lib==1.22.0
websockets==8.1
